% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{rabbitmq_close}
\alias{rabbitmq_close}
\title{Close an open connection to a rabbitmq host.}
\usage{
rabbitmq_close()
}
\value{
TRUE if the connection was closed or FALSE if the
        connection could not be closed.
}
\description{
This function will close the connection to a rabbitmq server. The
The function will return TRUE if the connection is closed, if there
is no connection, or the connection could not be closed it will
return FALSE.
}
\examples{
\dontrun{
rabbitmq_close()
}

}
