% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{rabbitmq_connect}
\alias{rabbitmq_connect}
\title{Connect to a rabbitmq host.}
\usage{
rabbitmq_connect(host = "localhost", port = 5672, vhost = "\%2F")
}
\arguments{
\item{host}{the hostname of the RabbitMQ server}

\item{port}{the hostname of the RabbitMQ server}

\item{vhost}{the hostname of the RabbitMQ server}
}
\value{
TRUE if the connection was created or FALSE if no
        connection could be created.
}
\description{
This function will connect to a rabbitmq server. The default host
it will connect to is localhost. This function needs to be called
once before all other functions are called. The function will
return TRUE if the connection is created successfully. If already
connected or the connection failed to create it will return FALSE.
}
\examples{
\dontrun{
stopifnot(rabbitmq_connect(host = "rabbitmq"))
}

}
